requirements.txt
streamlit
pandasg